test(docs): share architecture workspace fixtures
